reintroduce SparseAdam optimiser as LazyAdam (to avoid confusion)
