+++ test/test_optim.py
@@ -323,15 +323,6 @@ class TestOptim(TestCase):
         with self.assertRaisesRegex(ValueError, "Invalid beta parameter at index 0: 1.0"):
             optim.Adam(None, lr=1e-2, betas=(1.0, 0.0))
 
-    def test_adam_sparse(self):
-        self._test_rosenbrock_sparse(
-            lambda params: optim.Adam(params, lr=4e-2)
-        )
-        self._test_rosenbrock_sparse(
-            lambda params: optim.Adam(params, lr=4e-2),
-            [lambda opt: ReduceLROnPlateau(opt, threshold=1e-4)]
-        )
-
     def test_adamw(self):
         self._test_basic_cases(
             lambda weight, bias: optim.AdamW([weight, bias], lr=1e-3)
@@ -344,7 +335,7 @@ class TestOptim(TestCase):
 
     def test_sparse_adam(self):
         self._test_rosenbrock_sparse(
-            lambda params: optim.Adam(params, lr=4e-2),
+            lambda params: optim.SparseAdam(params, lr=4e-2),
             [],
             True
         )
